requirements.txt
flask==2.2.3
Werkzeug==2.2.3
Pillow==9.4.0
opencv-python==4.7.0.72
numpy==1.24.2
pytesseract==0.3.10
pdf2image==1.16.3
img2pdf==0.4.4
python-dotenv==1.0.0
scipy==1.10.1